Refactor FastUI page builders out of UI controller
